prim/Enum: Add another variant of SEAD_ENUM
There is a variant of SEAD_ENUM that seems to be used when no custom
enumerator values are required; it doesn't have findRelativeIndex
or ValueArray, and the relative index *is* the ValueType value.

Both variants are used.

This commit renames SEAD_ENUM (the complex variant) to SEAD_ENUM_EX
and implements the simpler variant as SEAD_ENUM.