Disposer: Fix several cases of UB and work around another UB issue
Do not reinterpret_cast as that is undefined behavior before the
placement new calls actually create objects.

Fix the placement new for sInstance: we must ensure the instance is
not zero initialised when CLASS has no user-provided constructor
because that will clobber the singleton disposer that was created
earlier (yikes). The whole thing reeks of UB and should probably be
replaced at some point.