thread/Atomic: Reimplement cmpxchg to match Nintendo's A64 version
Clang's atomic intrinsics cause an additional clrex to be emitted on
the failure path [1], which can cause basic blocks to be in the wrong
order in some cases. To avoid this situation, reimplement cmpxchg
manually using Clang's ARM specific intrinsics. Mark our version as
a matching hack because it is only useful for matching; the clrex is
useful for performance reasons.

[1] https://reviews.llvm.org/D13033