container/Buffer: Remove useless inline keyword to fix matching issues
In some situations, marking a function as inline can cause Clang to
decide *not* to inline it even when LLVM otherwise would inline it.
Totally counterproductive, so let's just remove it.

There are also several instances of tryAllocBuffer being inlined when
they shouldn't be in BotW. Removing inline fixes both issues.