Buffer: Reduce rebuild times by not including MathCalcCommon
The only reason we were including MathCalcCommon was to use Mathi::abs;
however, including MathCalcCommon means that every single user of
Buffer (and their downstream dependencies) needs to be rebuilt
whenever maths code changed. In BotW, this means ~3600 files to rebuild
every time a new function is added or fixed in MathCalcCommon.

To reduce rebuild times, we now ship with our own copy of Mathi::abs
and do not include MathCalcCommon anymore. This is not expected to be
a maintenance burden as the implementation of abs() is not expected
to ever change.