CalcCommon additions, minor fixes
MathCalcCommon::max was changed due to being the exact same implementation as ::clampMin, which is unnecessary, and the new implementation was needed.
QuatCalcCommon::setAxisAngle converts and axis and an angle into a quaternion, and a new Vector3CalcCommon::rotate was added for quaternion rotation on a vector. the implementation is matching in some functions, but has a register swap mismatch in others. it's as good as its gonna get for now.